pyproject.toml
[tool.poetry]
name = "strava-dash"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.18.1"
dash-table = "^5.0.0"
pandas = "^2.2.3"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
polyline = "^2.0.2"
folium = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
